[filmlist], [static], [templates] * added api for filmlist page * added js scripts for load additional content, checkbox toggling * reformatted html * added new styles * fully new style for filmlist admin page * moved sensetive data to local_settings.py * added settings.py back

+++ static/admin/js/form-actions.js
@@ -80,20 +80,24 @@ submitFilm = (form_id) => {
         });
 }
 
-removeFilm = (filmid) => {
-    let token = document.getElementsByName("csrfmiddlewaretoken")[0].value;
-    data = {
-        "csrfmiddlewaretoken": token,
-        "film_id": filmid,
+removeFilm = (filmid, film_name) => {
+    if (confirm(`Delete ${film_name}?`)) {
+        let token = document.getElementsByName("csrfmiddlewaretoken")[0].value;
+        data = {
+            "csrfmiddlewaretoken": token,
+            "film_id": filmid,
+        }
+        fetch("/admin/cinema/filmlist/", {
+            method: "DELETE",
+            headers: {
+                "X-CSRFToken": token,
+                "Accept": "application/json",
+                "Content-Type": "application/json"
+            },
+            body: JSON.stringify(data),
+        })
+            .then(response => getPage());
     }
-    fetch("/admin/cinema/filmlist/", {
-        method: "DELETE",
-        headers: {
-            "X-CSRFToken": token,
-            "Accept": "application/json",
-            "Content-Type": "application/json"
-        },
-        body: JSON.stringify(data),
-    })
-        .then(response => getPage());
 }
+
+
